Remove ProgressPill component, inline into ProgressLevelSet without Radium
